Move example tests to Examples namespace
