+++ src/Refitter.Tests/Examples/KebabCaseTests.cs
@@ -3,7 +3,7 @@ using Refitter.Core;
 using Refitter.Tests.Build;
 using Xunit;
 
-namespace Refitter.Tests;
+namespace Refitter.Tests.Examples;
 
 public class KebabCaseTests
 {
