+++ src/Refitter.Tests/Examples/PathParametersTests.cs
@@ -4,7 +4,7 @@ using Refitter.Core;
 using Refitter.Tests.Build;
 using Xunit;
 
-namespace Refitter.Tests;
+namespace Refitter.Tests.Examples;
 
 public class PathParametersTests
 {
